requirements.txt
Pillow>=10.0.0
jsonschema>=4.0.0
tkinterdnd2>=0.3.0
pytest>=7.0.0
